Split test_utils into its own module
